bye cmake

+++ src/ui/menu.c
@@ -1,5 +1,5 @@
 #include <menu.h>
 
-ui_menu_t* load_ui_menu(const char* title, ui_menu_element* elements, size_t element_count) {
+ui_menu_t* load_ui_menu(const char* title, ui_menu_element* elements, int element_count) {
     
 }
